requirements.txt
aniso8601==9.0.1
anyio==3.6.2
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
bcrypt==4.0.1
black==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorclass==2.2.2
comm==0.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-admin-interface==0.26.0
django-colorfield==0.9.0
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==23.2
django-mongodb-engine==0.6.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
djangotoolbox==1.8.0
dnspython==2.2.1
docopt==0.6.2
doctor==3.13.7
drf-yasg==1.21.7
executing==1.2.0
fastapi==0.88.0
filelock==3.12.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
flasks==0.1.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
inflection==0.5.1
ipython==8.13.1
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==2.6.0
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
motor==3.1.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
mysqlclient==2.2.0
nest-asyncio==1.5.6
openapi-codec==1.3.2
packaging==23.1
parso==0.8.3
pathspec==0.11.0
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip-autoremove==0.10.0
pip-upgrader==1.4.15
pipdeptree==2.11.0
pipenv==2022.11.30
pipreqs==0.4.13
platformdirs==3.8.0
ply==3.11
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
pyDes==2.0.1
Pygments==2.15.1
PyJWT==2.3.0
pymongo==4.3.3
PyMySQL==1.1.0
Pyomo==6.4.5.dev0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pyupgrade==3.9.0
PyYAML==6.0.1
pyzmq==25.0.2
razorpay==0.2.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
rfc3987==1.3.8
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.22.0
stevedore==5.1.0
strict-rfc3339==0.7
terminaltables==3.1.10
text-unidecode==1.3
tokenize-rt==5.1.0
tornado==6.3.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2021.5
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.2
whitenoise==6.5.0
yarg==0.1.9
